+++ core/math/src/main/java/org/arakhne/afc/math/MathInversableFunction.java
@@ -17,6 +17,7 @@
  */
 package org.arakhne.afc.math;
 
+
 /**
  * Define a mathematic inverse function.
  * 
@@ -33,6 +34,6 @@ public interface MathInversableFunction extends MathFunction {
 	 * @return the value of {@code f<sup>-1</sup>(y)}.
 	 * @throws MathException in case {@code f<sup>-1</sup>(y)} could not be computed
 	 */
-	public float inverseF(float y)  throws MathException;
+	public double inverseF(double y)  throws MathException;
 
 }
